moves test for get_tokenizer_from_nlu_config to tests/shared/utils/test_components.py
